API

Artifacts

PkgUtility.deploy_artifact! — Function

PkgUtility provides a convenient wrapper to deploy the artifact from a given list of files, for exmaple, deploy_artifact! function is used to deploy the GirddingMachine artifacts.

deploy_artifact!(art_toml, art_name, art_locf, art_file, art_tarf, art_urls; new_file)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/artifacts.jl:44.

deploy_artifact!(art_toml, art_name, art_locf, art_tarf, art_urls)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/artifacts.jl:100.

PkgUtility.deploy_artifact! — Method

What deploy_artifact! function does are

determine if the artifact already exists in the art_toml file

if true, skip the deployment

if false

copy the file(s) to ~/.julia/artifacts/ARTIFACT_SHA/

compress the artifact file(s) to a .tar.gz file

calculate the hash value of the compressed tar.gz file

bind the artifact file to the .toml file

Method for this deployment is

deploy_artifact!(art_toml::String, art_name::String, art_locf::String, art_file::Vector{String}, art_tarf::String, art_urls::Vector{String}; new_file::Vector{String} = art_file)

Deploy the artifact, given

art_toml Artifact .toml file location

art_name Artifact name identitfier

art_locf Local folder that stores the source files

art_file Vector of the source file names

art_tarf Folder location to store the compressed .tar.gz file

art_urls Vector of public urls, where the compressed files are to be uploaded (user need to upload the file manually)

new_file Optional. New file names of the copied files (same as art_file by default)

Examples

# deploy art_1.txt and art_2.txt as test_art artifact
deploy_artifact!("Artifacts.toml", "test_art", "./", ["art_1.txt", "art_2.txt], "./", ["https://public.server.url"]);

# deploy art_1.txt and art_2.txt as test_art artifact with new names
deploy_artifact!("Artifacts.toml", "test_art", "./", ["art_1.txt", "art_2.txt], "./", ["https://public.server.url"]; new_files=["new_1.txt", "new_2.txt"]);

PkgUtility.deploy_artifact! — Method

In many cases, one might want to copy all the files in a folder to the target artifact, and iterate the file names is not convenient at all. Thus, a readily usable method is provided for this purpose:

deploy_artifact!(art_toml::String, art_name::String, art_locf::String, art_tarf::String, art_urls::Vector{String})

Deploy the artifact, given

art_toml Artifact .toml file location

art_name Artifact name identitfier

art_locf Local folder that stores the source files (all files will be copied into the artifact)

art_tarf Folder location to store the compressed .tar.gz file

art_urls Vector of public urls, where the compressed files are to be uploaded (user need to upload the file manually)

Examples

# deploy all files in target folder
deploy_artifact!("Artifacts.toml", "test_art", "./folder", "./", ["https://public.server.url"]);

Date

PkgUtility.parse_timestamp — Function

Researchers may often run into the scenarios when they have to convert dates to strings or numbers, vice versa. Thought the conversions are easy to make, it is not convenient at all if ones need to recreate the wheel everytime. Thus, we provide functions to make the conversion. One of the common conversions is to convert time stamp to a number:

parse_timestamp(time_stamp; in_format, out_format)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/date/doy.jl:85.

parse_timestamp(year, doy)
parse_timestamp(year, doy, sep)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/date/doy.jl:153.

PkgUtility.parse_timestamp — Method

The time stamp can be either an integer or a string, and the integer will be converted to a string within the function. The supported time stamp formats are YYYYMMDD, YYYYMMDDhh, YYYYMMDDhhmm, and YYYYMMDDhhmmss, where the labels are

YYYY Year number

MM Month number

DD Day number

hh Hour number

mm Minute number

ss second number

The supported outputs are

DATE A Dates.Date type variable

DATETIME A Dates.DateTime type variable

DOY A day of year integer

FDOY A day of year float

parsetimestamp( timestamp::Union{Int,String}; informat::String = "YYYYMMDD", outformat::String = "DOY")

Convert time stamp, given

time_stamp Time stamp

in_format Format of time_stamp, default is YYYYMMDD

out_format Output format, default is DOY

Examples

# if input format is YYYYMMDD
_inf = "YYYYMMDD";
for _ts in [20200130, "20200130"]
    for _outf in ["DATE", "DATETIME", "DOY", "FDOY"]
        time = parse_timestamp(_ts; in_format=_inf, out_format=_outf);
    end
end

# if input format is YYYYMMDDhh
_inf = "YYYYMMDDhh";
for _ts in [2020013002, "2020013002"]
    for _outf in ["DATE", "DATETIME", "DOY", "FDOY"]
        time = parse_timestamp(_ts; in_format=_inf, out_format=_outf);
    end
end

# if input format is YYYYMMDDhhmm
_inf = "YYYYMMDDhhmm";
for _ts in [202001300215, "202001300215"]
    for _outf in ["DATE", "DATETIME", "DOY", "FDOY"]
        time = parse_timestamp(_ts; in_format=_inf, out_format=_outf);
    end
end

# if input format is YYYYMMDDhhss
_inf = "YYYYMMDDhhmmss";
for _ts in [20200130021515, "20200130021515"]
    for _outf in ["DATE", "DATETIME", "DOY", "FDOY"]
        time = parse_timestamp(_ts; in_format=_inf, out_format=_outf);
    end
end

PkgUtility.parse_timestamp — Function

Also, one might want to convert date to time stamp. Supported conversions are

parse_timestamp(year::Int, doy::Int, sep::String="")

Convert date to YYYY(sep)MM(sep)DD format, given

year Year

doy Day of year (typically 1-365, 1-366 for leap years)

sep String that separates year, month, and day. Default is ""

Examples

# convert the time to YYYYMMDD format
time = parse_timestamp(2020, 100);
time = parse_timestamp(2020, 100, "");

# convert the time to YYYY.MM.DD format
time = parse_timestamp(2020, 100, ".");

PkgUtility.month_days — Function

month_days(year::Int, month::Int)

How many days in a month, given

year Year

month Month

Examples

# inquire how many days in 2020 October
mday = month_days(2020, 10);

PkgUtility.month_ind — Function

month_ind(year::Int, doy::Int)

Which month does the day of year belongs to, given

year Year

doy Day of year (typically 1-365, 1-366 for leap years)

Examples

# inquire which month it is for 100th day in 2020
mind = month_ind(2020, 100);

PkgUtility.terror — Function

A function to covert information to time + information

terror(info::String)

Display the error info with date and time before the given

info Infomation to display

Examples

# add a time stamp before the message
@error terror("This is an error with time stamp at the beginning!");

PkgUtility.tinfo — Function

A function to covert information to time + information

tinfo(info::String)

Display the error info with date and time before the given

info Infomation to display

Examples

# add a time stamp before the message
@info tinfo("This is an info with time stamp at the beginning!");

PkgUtility.twarn — Function

A function to covert information to time + information

twarn(info::String)

Display the error info with date and time before the given

info Infomation to display

Examples

# add a time stamp before the message
@warn twarn("This is an info with time stamp at the beginning!");

Display

PkgUtility.pretty_display! — Function

Display the parameters in a colored way.

pretty_display!(dict, max_len)
pretty_display!(dict, max_len, spaces)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/display/recursive.jl:36.

pretty_display!(dicts)
pretty_display!(dicts, spaces)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/display/recursive.jl:97.

PkgUtility.pretty_display! — Function

When a dictionary (a pair of key and item) is passed to the pretty_display! function, the dictionary will be displayed as (LEADING_SPACE)key ⇨ item,. However, if the dictionary item is an array of dictionaries, then the array will be displayed recursively.

pretty_display!(dict::Pair, max_len::Int, spaces = "    ")

Display the key and item in a dictionary (Pair), given

dict Dictionary to display

max_len Maximum length of dictionary key (in an array)

spaces Leading spaces before displaying the dictionary key

Examples

# display the pair in a pretty way (automatic alignments)
pretty_display!("a" => "b", 1);
pretty_display!("a" => "b", 2, "  ");

PkgUtility.pretty_display! — Function

When an array of dictionaries is given, the pretty_display! function computes the maximum length of the dictionary keys, and then display the dictionaries in a colored and pretty manner:

pretty_display!(
            dicts::Union{Vector{Pair{String,String}},
                         Vector{Pair{String,Any}},
                         Vector{Pair{Any,String}},
                         Vector{Pair{Any,Any}}},
            spaces::String = "    ")

Display array of pairs (dictionary) in recursive manner, given

dicts Parameters to display

spaces Leading spaces

Examples

# display the vector of pairs recursively (automatic alignments)
_dicts = ["A" => "b", "d" => "A", "rr" => ["ra" => "rB", "rD" => "ra"]];
pretty_display!(_dicts);
pretty_display!(_dicts, "  ");

IO

CSV

PkgUtility.read_csv — Function

Reading CSV file may be troublesome for many cases, for example, when some lines exist before the header or when a few unit lines exist after the header. Here we provide a generalized function to read CSV files:

read_csv(file; skiprows, displaying)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/io/csv.jl:38.

PkgUtility.read_csv — Method

When there is no unit row in the CSV file, the method is given as

read_csv(file::String; skiprows::Int = 0, displaying::Bool = false)

Read CSV file as a DataFrame, given

file Path to CSV file

skiprows Rows to skip

Example

# read CSV without skipping any rows
df = read_csv("test.csv");
df = read_csv("test.csv"; skiprows=0, displaying=true);

# skip first 2 rows
df = read_csv("test.csv"; skiprows=2, displaying=true);

PkgUtility.save_csv! — Function

Saving data to CSV file may also be troublesome when we want to save the data using a known format, for example, when we want to save attributes in the file as well such as the units. To best use Julia, we provide a generalized function to save data as CSV:

save_csv!(df, file; displaying)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/io/csv.jl:89.

save_csv!(file, df; displaying)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/io/csv.jl:100.

save_csv!(file, data, var_names; per_row)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/io/csv.jl:149.

save_csv!(data, var_names, file; per_row)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/io/csv.jl:194.

save_csv!(file, data; per_row)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/io/csv.jl:225.

save_csv!(data, file; per_row)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/io/csv.jl:231.

PkgUtility.save_csv! — Method

To save the data as a plain CSV without any attribute information, one may use one of these functions. Note that the sequence of the data frame and file name does not matter as the function will automatically recognize the sequence:

save_csv!(df::DataFrame, file::String; displaying::Bool = false)
save_csv!(file::String, df::DataFrame; displaying::Bool = false)

Save data to CSV file, given

df A DataFrame

file Path of the target CSV file

Example

# save data to CSV without any rows to skip
df = read_csv("in.csv"; skiprows=2);
save_csv!(df, "out.csv");
save_csv!("out.csv", df);

PkgUtility.save_csv! — Method

In many cases, one might have the data as Vector or Matrix, for example, the result from multiple threading calculations using pmap. In this case, converting the data to a DataFrame and then save the DataFrame may be the best case here. However, the results from pmap may be a Vector (R) of Vector (N), or a Vector (R) of Matrix (1×N), or a Vector (R) of Tuple, or a Vector of combinations. While it is possible to use vcat or hcat to recast the results to certain format, one needs to be really cautious about the recasting. Here we provide a wrapper function to automatically detect the result format and save the data. Again, one may reverse the order of file name and data in the function parameter list:

save_csv!(file::String,
          data::Vector,
          var_names::Vector{String};
          per_row::Bool=true)
save_csv!(data::Vector,
          var_names::Vector{String},
          file::String;
          per_row::Bool=true)

Save data to CSV file, given

file Path of the target CSV file

data A Vector of data, needs to be a Vector of Tuple (N), Vector (N), or Matrix (1×N)

var_name Variable names appear in the CSV file

per_row If true, item in the data Vector is treated as row; othrewise, the data item is treated as column

Example

# when data has equal length
save_csv!("test.csv", [(1,2), (1,2)], ["A", "B"]);
save_csv!("test.csv", [[1,2], [1,2]], ["A", "B"]);
save_csv!("test.csv", [[1 2], [1 2]], ["A", "B"]);
save_csv!("test.csv", [[1,2], [1 2], (1,2)], ["A", "B"]);

# when data has unequal length
save_csv!("test.csv", [(1,2), [1 2 3], [1,2,3,4]], ["A", "B"]);
save_csv!("test.csv", [(1,2), [1 2 3], [1,2,3,4]], ["A", "B", "C", "D", "E"]);

# use data item as column rather than row (default)
save_csv!("test.csv", [[1 2], [1 2]], ["A", "B"]; per_row=false);
save_csv!("test.csv", [[1 2], [1 2 3]], ["A", "B"]; per_row=false);

PkgUtility.save_csv! — Method

Further, as the above method supports filling in missing values, it is possible to not set the var_names:

save_csv!(file::String, data::Vector; per_row::Bool=true)
save_csv!(data::Vector, file::String; per_row::Bool=true)

Save data to CSV file, given

file Path of the target CSV file

data A Vector of data, needs to be a Vector of Tuple (N), Vector (N), or Matrix (1×N)

per_row If true, item in the data Vector is treated as row; othrewise, the data item is treated as column

Example

# when data has equal length
save_csv!("test.csv", [[1,2], [1 2], (1,2)]);

# when data has unequal length
save_csv!("test.csv", [(1,2), [1 2 3], [1,2,3,4]]);

# use data item as column rather than row (default)
save_csv!("test.csv", [[1 2], [1 2]]; per_row=false);
save_csv!("test.csv", [[1 2], [1 2 3]]; per_row=false);

DataFrame

PkgUtility.dataframe — Function

Create a data frame from given parameters:

dataframe()

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/io/dataframe.jl:30.

dataframe(v_data, v_name)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/io/dataframe.jl:46.

PkgUtility.dataframe — Method

If no parameter is passed to dataframe, an empty DataFrame will be returned:

dataframe()

Return an empty DataFrame

Example

df = dataframe();

PkgUtility.dataframe — Method

When a vector of data and a vector of headers are given, dataframe returns a DataFrame with the given data and header:

dataframe(v_data::Vector, v_name::Vector{String})

Return a DataFrame for given

v_data Vector of data

v_name Vector of data header

Note that the data and header vectors must have the same length.

Email

PkgUtility.send_email! — Function

Running program on the server mat take very long time, and one may not know the status without logging into the system. Thus, we provide a function to send email to notify the status:

send_email!(subject, from_email, to_email, body)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/io/email.jl:43.

PkgUtility.send_email! — Method

Here is the function to send email out:

send_email!(subject::String,
            from_email::String,
            to_email::String,
            body::String)

Send out email, given

subject Email subject

from_email The outgoing email address

to_email Email address to send out

body Main body of the email

Note that you have to run this function on a machine with sendmail set up.

Example

send_email!("[DO NOT REPLY] Job status", "from.email", "to.email", "This is the main text");

Land Wrapper

PkgUtility.AVOGADRO — Function

Avogadro's number [molecule mol⁻¹]

PkgUtility.CP_D — Function

Isobaric specific heat of dry air [J kg⁻¹ K⁻¹]

PkgUtility.CP_D_MOL — Function

Isobaric specific heat of dry air [J mol⁻¹ K⁻¹]

PkgUtility.CP_L — Function

Isobaric specific heat of liquid water [J kg⁻¹ K⁻¹]

PkgUtility.CP_V — Function

Isobaric specific heat of water vapor [J kg⁻¹ K⁻¹]

PkgUtility.GAS_R — Function

Universal gas constant [J mol⁻¹ K⁻¹]

PkgUtility.GRAVITY — Function

Gravity of the Earth [m s⁻²]

PkgUtility.H_PLANCK — Function

Planck constant [m² kg s⁻¹]

PkgUtility.K_BOLTZMANN — Function

Boltzmann constant [m² kg s⁻² K⁻¹]

PkgUtility.K_STEFAN — Function

Stefan-Boltzmann constant [W m⁻² K⁻⁴]

PkgUtility.K_VON_KARMAN — Function

Von Karman constant [-]

PkgUtility.LH_V0 — Function

Latent heat vaporization at $T_0$ [K kg⁻¹]

PkgUtility.LIGHT_SPEED — Function

Light speed in vacuum [m s⁻¹]

PkgUtility.M_DRYAIR — Function

Molar mass of dry air [kg mol⁻¹]

PkgUtility.M_H₂O — Function

Molar mass of water [kg mol⁻¹]

PkgUtility.P_ATM — Function

Mean atmospheric pressure at sea level [Pa]

PkgUtility.PRESS_TRIPLE — Function

Water vapor pressure at triple temperature [Pa]

PkgUtility.R_V — Function

Gas constant water vapor [J kg⁻¹ K⁻¹]

PkgUtility.RT_25 — Function

Gas constant times 298.15 K [J mol⁻¹]

PkgUtility.T_0 — Function

Freezing temperature of water [K]

PkgUtility.T_25 — Function

Kelvin temperature at 25 Celcius [K]

PkgUtility.T_TRIPLE — Function

Triple temperature of water [K]

PkgUtility.V_H₂O — Function

Molar volume of liqiud water

PkgUtility.YEAR_D — Function

Mean number of days per year [day]

PkgUtility.ρ_H₂O — Function

Density of liquid water [kg m⁻³]

PkgUtility.ρg_MPa — Function

Density of water times gravity [MPa m⁻¹]

Math

Integral

PkgUtility.numerical∫ — Function

Function numerical∫ calculates the integral numerically. The supported methods are

numerical∫(f, Δx)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/math/integral.jl:36.

numerical∫(f, Δx)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/math/integral.jl:67.

numerical∫(f, x_min, x_max, n)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/math/integral.jl:102.

numerical∫(f, x_min, x_max)
numerical∫(f, x_min, x_max, x_tol)
numerical∫(f, x_min, x_max, x_tol, y_tol)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/math/integral.jl:148.

PkgUtility.numerical∫ — Method

When two arrays (one for f(x) and one for Δx) are provided, numerical∫ computes the sum of the mutiplication of the two using matrix mutiplication fxs' * Δxs.

numerical∫(f::Array{FT,1}, Δx::Array{FT,1}) where {FT<:AbstractFloat}

Intergal of given

f f(x) for each x

Δx Δx for each x

Note that f and Δx may have different dimensions, and if so a warning will display.

Examples

FT = Float32;
f_sum = numerical∫(FT[1,2,3,4], FT[0.1,0.1,0.2,0.3]);

PkgUtility.numerical∫ — Method

The above methods is useful for both evenly and non-evenly distributed Δx. However, for many cases Δx is evenly distributed and provided as a number rather than an array. In this scenario, a special method is given:

numerical∫(f::Array{FT,1}, Δx::FT) where {FT<:AbstractFloat}

Intergal of given

f f(x) for each x

Δx Δx for x

Examples

FT = Float32;
f_sum = numerical∫(FT[1,2,3,4], FT(0.1));

PkgUtility.numerical∫ — Method

We also provide a function to manually solve for the integral of a given function for a given range of x. The method is

numerical∫(f::Function,
           x_min::FT,
           x_max::FT,
           n::Int
) where {FT<:AbstractFloat}

Intergal of given

f A function

x_min Minimum limit of x

x_max Maximum limit of x

n Number of points in the x range (evenly stepped)

Examples

FT = Float32;
func(x) = x^2;
f_sum = numerical∫(func, FT(0), FT(2), 20);

PkgUtility.numerical∫ — Function

This method automatically computes the integral of a function for an x within a range:

numerical∫(f::Function,
           x_min::FT,
           x_max::FT,
           x_tol::FT = sqrt(eps(FT)),
           y_tol::FT = sqrt(eps(FT))
) where {FT<:AbstractFloat}

Intergal of given

f A function

x_min Minimum limit of x

x_max Maximum limit of x

x_tol Tolerance of Δx (x/N)

y_tol Tolerance of the integral solution

Example

FT = Float32;
func(x) = x^2;
f_sum = numerical∫(func, FT(0), FT(2));
f_sum = numerical∫(func, FT(0), FT(2), FT(1e-3), FT(1e-3));

Quadratic solver

PkgUtility.lower_quadratic — Function

lower_quadratic(a::FT, b::FT, c::FT) where {FT<:AbstractFloat}

Return the lower quadratic solution or NaN, given

a Parameter in a*x^2 + b*x + c = 0

b Parameter in a*x^2 + b*x + c = 0

c Parameter in a*x^2 + b*x + c = 0

Example

low_x = lower_quadratic(1.0, 2.0, 0.0);
low_x = lower_quadratic(-1.0, 2.0, 0.0);
low_x = lower_quadratic(1.0, 2.0, 2.0);

PkgUtility.upper_quadratic — Function

upper_quadratic(a::FT, b::FT, c::FT) where {FT<:AbstractFloat}

Return the upper quadratic solution or NaN, given

a Parameter in a*x^2 + b*x + c = 0

b Parameter in a*x^2 + b*x + c = 0

c Parameter in a*x^2 + b*x + c = 0

Example

up_x = upper_quadratic(1.0, 2.0, 0.0);
up_x = upper_quadratic(-1.0, 2.0, 0.0);
up_x = upper_quadratic(1.0, 2.0, 2.0);

Statistics extensions

PkgUtility.nanmax — Function

nanmax(x::Array)

Return the maximum of array ommiting the NaN, given

x Array of numbers, can be NaN

Example

xs = [1, 2, 4, NaN];
nmax = nanmax(xs);

PkgUtility.nanmean — Function

nanmean(x::Array)

Return the mean of array by ommiting the NaN, given

x Array of numbers, can be NaN

Example

xs = [1, 2, 4, NaN];
nmean = nanmean(xs);

PkgUtility.nanmedian — Function

nanmedian(x::Array)

Return the median of array by ommiting the NaN, given

x Array of numbers, can be NaN

Example

xs = [1, 2, 4, NaN];
nmed = nanmedian(xs);

PkgUtility.nanmin — Function

nanmin(x::Array)

Return the maximum of array ommiting the NaN, given

x Array of numbers, can be NaN

Example

xs = [1, 2, 4, NaN];
nmin = nanmin(xs);

PkgUtility.nanpercentile — Function

nanpercentile(x::Array, p::Number)

Return the percentile by excluding the NaN of given

x Array of data

p Percentile

Example

xs = rand(100);
pth = nanpercentile(rand(100), 50);
xs[1:10] .= NaN;
pth = nanpercentile(rand(100), 50);

PkgUtility.nanstd — Function

nanstd(x::Array)

Return the std of array by ommiting the NaN, given

x Array of numbers, can be NaN

Example

xs = [1, 2, 4, NaN];
nstd = nanstd(xs);

PkgUtility.mae — Function

mae(y::Array, pred::Array)

Return the mean absolute error by ommiting the NaN, given

y Array of numbers, can be NaN

pred Array of predictions, can be NaN

Example

ys = [1, 2, 4, NaN];
preds = [2, 4, 1, NaN];
nmae = mae(ys, preds);

PkgUtility.mape — Function

mape(y::Array, pred::Array)

Return the mean absolute percentage error by ommiting the NaN, given

y Array of numbers, can be NaN

pred Array of predictions, can be NaN

Example

ys = [1, 2, 4, NaN];
preds = [2, 4, 1, NaN];
nmape = mape(ys, preds);

PkgUtility.mase — Function

mase(y::Array, pred::Array)

Return the mean absolute standardized error by ommiting the NaN, given

y Array of numbers, can be NaN

pred Array of predictions, can be NaN

Example

ys = [1, 2, 4, NaN];
preds = [2, 4, 1, NaN];
nmase = mase(ys, preds);

PkgUtility.rmse — Function

rmse(y::Array, pred::Array)

Return the root mean square error by ommiting the NaN, given

y Array of numbers, can be NaN

pred Array of predictions, can be NaN

Example

ys = [1, 2, 4, NaN];
preds = [2, 4, 1, NaN];
rmse = rmse(ys, preds);

Recursive test

PkgUtility.FT_test — Function

Using consistent floating number type may accelerate calculations, particularly for GPUs. Thus, we provided a function to test the floating number type recursively:

FT_test(para, FT)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/test/recursive.jl:37.

FT_test(para, FT)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/test/recursive.jl:76.

FT_test(para)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/test/recursive.jl:114.

FT_test(para, FT)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/test/recursive.jl:144.

PkgUtility.FT_test — Method

If the testing variable is an array, the function will test if element type is float number:

If true, the function tests if the element type is the same as given FT

If false, the function tests each element recursively

FT_test(para::Array, FT)

Test the the floating point type of a variable, given

para An array

FT given FT

Example

ft_1 = FT_test([1, 2, 3], Float64);
ft_2 = FT_test(Any[1, 1.0f0, 1.0e0], Float64);
ft_3 = FT_test([1, 2.0, "a"], Float64);

PkgUtility.FT_test — Method

When the given variable is a number, the following criteria are used based on the type of the number

If the variable is a float, then compare it to FT

Otherwise, return true (because integers can be mix-used with floats)

FT_test(para::Number, FT)

Test the the floating point type of a variable, given

para A number

FT given FT

Example

ft_1 = FT_test(1, Float64);
ft_2 = FT_test(1.0, Float64);
ft_3 = FT_test(1.0f0, Float64);

PkgUtility.FT_test — Method

If the variable is a function, module, or symbol, then nothing will be done, and a true will be returned. This method is to avoid endless nested loop within the test of Function or Module:

FT_test(para::Union{Function,Module,Symbol})

Test the the floating point type of a variable, given

para An array

FT given FT

Example

f(x) = x^2;
m = CLIMAParameters;

ft_1 = FT_test(f, Float64);
ft_2 = FT_test(m, Float64);
ft_3 = FT_test("haha", Float64);

PkgUtility.FT_test — Method

The variable to test maybe a struct, but FT_test does not know the struct type name a priori. Thus, we try to read out the fields of the variable:

If succeeds, the function test the fields recursively

If fails, then do nothing

FT_test(para::Any, FT)

Test the the floating point type of a variable, given

para A function, module, or symbol

FT given FT

Example

struct SA
    a
    b
end
sa = SA(1, 2.0);

ft_1 = FT_test(sa);
ft_2 = FT_test();

PkgUtility.NaN_test — Function

Like FT_test, same logic is used to test if all the elements within the tested variable are not NaN:

NaN_test(para)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/test/recursive.jl:208.

NaN_test(para)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/test/recursive.jl:244.

NaN_test(para)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/test/recursive.jl:269.

NaN_test(para)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/test/recursive.jl:294.

PkgUtility.NaN_test — Method

When an array is passed to the function, the function first determines if the variable is an array of number:

If true, the function tests each element within the array directly

If false, the function tests each element recursively

NaN_test(para::Array)

Test the the floating point type of para is not NaN, given

para An array

Example

nan_1 = NaN_test([1,2,NaN]);
nan_2 = NaN_test([1,3,4]);
nan_3 = NaN_test([1,2,"a"]);

PkgUtility.NaN_test — Method

If the variable is a number, the function tests if the number is NaN directly:

NaN_test(para::Number)

Test if the variable is NaN, given

para A number

Example

nan_1 = NaN_test(1);
nan_2 = NaN_test(NaN);

PkgUtility.NaN_test — Method

If the variable is a function, module, or symbol, then nothing will be done, and a true will be returned. This method is to avoid endless nested loop within the test of Function or Module:

NaN_test(para::Union{Function,Module,Symbol})

Test the the floating point type of a variable, given

para A function, module, or symbol

Example

f(x) = x^2;
m = CLIMAParameters;

nan_1 = NaN_test(f);
nan_2 = NaN_test(m);
nan_3 = NaN_test("haha");

PkgUtility.NaN_test — Method

If the given variable is a struct, the function will test the fields recursively:

NaN_test(para::Any)

Test if the variable contains any NaN, given

para Tested variable

Example

struct SA
    a
    b
end

nan_1 = NaN_test(SA(1,2));
nan_2 = NaN_test(SA(1,NaN));