API

Artifacts

PkgUtility.deploy_artifact — Function

deploy_artifact(
            art_toml::String,
            art_name::String,
            art_locf::String,
            art_file::Array{String,1},
            art_tarf::String,
            art_urls::Array{String,1};
            new_file::Array{String,1} = art_file)
deploy_artifact(
            art_toml::String,
            art_name::String,
            art_locf::String,
            art_tarf::String,
            art_urls::Array{String,1})

Deploy the artifact, given

art_toml Artifact TOML file location

art_name Artifact name identitfier

art_locf Local folder that stores artifact files

art_file Array of artifact file names

art_tarf Folder location to store artifact TAR file

art_urls Array of urls to upload the Artifact to

new_file Optional. New file names (same as art_file by default)

PkgUtility.predownload_artifact — Function

predownload_artifact(name::String, artifact_toml::String)

Download the artifact from given server if it does not exist, given

name Artifact name

artifact_toml Artifacts.toml file location

Date

PkgUtility.doy_to_int — Function

doy_to_int(year::Int, doy::Int)

Convert date to YYYYMMDD format to work with flux tower datasets, given

year Year with the format of XXXX

doy Day of year (typically 1-365, 1-366 for leap years)

PkgUtility.int_to_doy — Function

int_to_doy(timestamp::String)

Convert date from YYYYMMDD to day of year, given

timestamp Time stamp as in flux tower dataset

Display

PkgUtility.pretty_display — Function

pretty_display(
            para::Array{Pair{Union{Any},Union{Any}},1},
            spaces::String = " ")

Display array of pairs (dictionary) in recursive manner, given

para Parameters to display

spaces Leading spaces

Recursive test

PkgUtility.FT_test — Function

FT_test(para::Any, FT)

Test the the floating point type of para is FT, given

para Any type of parameters

FT given FT

If data type is not supported, use pass!

PkgUtility.NaN_test — Function

NaN_test(para::Any)

Test the the floating point type of para is not NaN, given

para Any type of parameters

If data type is not supported, use pass!