API

Artifacts

PkgUtility.deploy_artifact — Function

deploy_artifact(
            art_toml::String,
            art_name::String,
            art_locf::String,
            art_file::Array{String,1},
            art_tarf::String,
            art_urls::Array{String,1};
            new_file::Array{String,1} = art_file)
deploy_artifact(
            art_toml::String,
            art_name::String,
            art_locf::String,
            art_tarf::String,
            art_urls::Array{String,1})

Deploy the artifact, given

art_toml Artifact TOML file location

art_name Artifact name identitfier

art_locf Local folder that stores artifact files

art_file Array of artifact file names

art_tarf Folder location to store artifact TAR file

art_urls Array of urls to upload the Artifact to

new_file Optional. New file names (same as art_file by default)

PkgUtility.predownload_artifact — Function

predownload_artifact(name::String, artifact_toml::String)

Download the artifact from given server if it does not exist, given

name Artifact name

artifact_toml Artifacts.toml file location

Date

PkgUtility.int_to_doy — Function

int_to_doy(timestamp::String)

Convert date from YYYYMMDD to day of year, given

timestamp Time stamp as in flux tower dataset

PkgUtility.month_days — Function

month_days(year::Int, month::Int)

Return how many days for given

year Year

month Month

PkgUtility.month_ind — Function

month_ind(year::Int, doy::Int)

Return which month does the day belongs to

year Year

doy Day of year (typically 1-365, 1-366 for leap years)

PkgUtility.parse_date — Function

parse_date(year::Int, doy::Int, sep::String="")

Convert date to YYYY(sep)MM(sep)DD format to work with flux tower datasets, given

year Year

doy Day of year (typically 1-365, 1-366 for leap years)

sep String that separates year, month, and day. Default is ""

Display

PkgUtility.pretty_display — Function

pretty_display(
            para::Array{Pair{Union{Any},Union{Any}},1},
            spaces::String = " ")

Display array of pairs (dictionary) in recursive manner, given

para Parameters to display

spaces Leading spaces

Math

Integral function

PkgUtility.numerical∫ — Function

numerical∫(f::Array{FT,1}, dx::Array{FT,1}) where {FT<:AbstractFloat}
numerical∫(f::Array{FT,1}, dx::FT) where {FT<:AbstractFloat}

A fast way of integrating functions, given

f f(x) for each x

dx Delta x for each x

Note that f and dx must have the same length

Quadratic solver

PkgUtility.lower_quadratic — Function

lower_quadratic(a::FT, b::FT, c::FT) where {FT<:AbstractFloat}

Return the lower quadratic solution or NaN, given

a Parameter in a*x^2 + b*x + c = 0

b Parameter in a*x^2 + b*x + c = 0

c Parameter in a*x^2 + b*x + c = 0

PkgUtility.upper_quadratic — Function

upper_quadratic(a::FT, b::FT, c::FT) where {FT<:AbstractFloat}

Return the upper quadratic solution or NaN, given

a Parameter in a*x^2 + b*x + c = 0

b Parameter in a*x^2 + b*x + c = 0

c Parameter in a*x^2 + b*x + c = 0

Statistics extensions

PkgUtility.mae — Function

mae(y::Array, pred::Array)

Return the mean absolute error by ommiting the NaN, given

y Array of numbers, can be NaN

pred Array of predictions, can be NaN

PkgUtility.mape — Function

mape(y::Array, pred::Array)

Return the mean absolute percentage error by ommiting the NaN, given

y Array of numbers, can be NaN

pred Array of predictions, can be NaN

PkgUtility.mase — Function

mase(y::Array, pred::Array)

Return the mean absolute standardized error by ommiting the NaN, given

y Array of numbers, can be NaN

pred Array of predictions, can be NaN

PkgUtility.nanmax — Function

nanmax(x::Array)

Return the maximum of array ommiting the NaN, given

x Array of numbers, can be NaN

PkgUtility.nanmean — Function

nanmean(x::Array)

Return the mean of array by ommiting the NaN, given

x Array of numbers, can be NaN

PkgUtility.nanmedian — Function

nanmedian(x::Array)

Return the median of array by ommiting the NaN, given

x Array of numbers, can be NaN

PkgUtility.nanmin — Function

nanmin(x::Array)

Return the maximum of array ommiting the NaN, given

x Array of numbers, can be NaN

PkgUtility.nanstd — Function

nanstd(x::Array)

Return the std of array by ommiting the NaN, given

x Array of numbers, can be NaN

PkgUtility.rmse — Function

rmse(y::Array, pred::Array)

Return the root mean square error by ommiting the NaN, given

y Array of numbers, can be NaN

pred Array of predictions, can be NaN

PkgUtility.ncread — Function

ncread(FT, file::String, var::String)

Read data from nc file, given

FT Float number type

file Dataset path

var Variable name

Note that this ncread function differs from the NetCDF.ncread in that it accounts for scale_factor and add_offset in the attributes if they exist.

Recursive test

PkgUtility.FT_test — Function

FT_test(para::Any, FT)

Test the the floating point type of para is FT, given

para Any type of parameters

FT given FT

If data type is not supported, use pass!

PkgUtility.NaN_test — Function

NaN_test(para::Any)

Test the the floating point type of para is not NaN, given

para Any type of parameters

If data type is not supported, use pass!