API

Artifacts

PkgUtility.deploy_artifact — Function

PkgUtility provides a convenient wrapper to deploy the artifact from a given list of files, for exmaple, deploy_artifact function is used to deploy the GirddingMachine artifacts.

deploy_artifact(art_toml, art_name, art_locf, art_file, art_tarf, art_urls; new_file)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/artifact/deploy.jl:71.

deploy_artifact(art_toml, art_name, art_locf, art_tarf, art_urls)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/artifact/deploy.jl:146.

PkgUtility.deploy_artifact — Method

What deploy_artifact function does are

determine if the artifact already exists in the art_toml file

if true, skip the deployment

if false

copy the file(s) to ~/.julia/artifacts/ARTIFACT_SHA/

compress the artifact file(s) to a .tar.gz file

calculate the hash value of the compressed tar.gz file

bind the artifact file to the .toml file

Method for this deployment is

deploy_artifact(
            art_toml::String,
            art_name::String,
            art_locf::String,
            art_file::Vector{String},
            art_tarf::String,
            art_urls::Vector{String};
            new_file::Vector{String} = art_file)

Deploy the artifact, given

art_toml Artifact .toml file location

art_name Artifact name identitfier

art_locf Local folder that stores the source files

art_file Vector of the source file names

art_tarf Folder location to store the compressed .tar.gz file

art_urls Vector of public urls, where the compressed files are to be uploaded (user need to upload the file manually)

new_file Optional. New file names of the copied files (same as art_file by default)

Examples

# deploy art_1.txt and art_2.txt as test_art artifact
deploy_artifact("Artifacts.toml", "test_art", "./", ["art_1.txt", "art_2.txt],
                "./", ["https://public.server.url"])
# deploy art_1.txt and art_2.txt as test_art artifact with new names
deploy_artifact("Artifacts.toml", "test_art", "./", ["art_1.txt", "art_2.txt],
                "./", ["https://public.server.url"];
                new_files=["new_1.txt", "new_2.txt"]);

PkgUtility.deploy_artifact — Method

In many cases, one might want to copy all the files in a folder to the target artifact, and iterate the file names is not convenient at all. Thus, a readily usable method is provided for this purpose:

deploy_artifact(
            art_toml::String,
            art_name::String,
            art_locf::String,
            art_tarf::String,
            art_urls::Vector{String})

Deploy the artifact, given

art_toml Artifact .toml file location

art_name Artifact name identitfier

art_locf Local folder that stores the source files (all files will be copied into the artifact)

art_tarf Folder location to store the compressed .tar.gz file

art_urls Vector of public urls, where the compressed files are to be uploaded (user need to upload the file manually)

Examples

# deploy all files in target folder
deploy_artifact("Artifacts.toml", "test_art", "./folder", "./",
                ["https://public.server.url"])

PkgUtility.predownload_artifact — Function

Artifacts stored on FTP cannot be installed directly for Windows using Julia 1.5. This problem may be fixed in future Julia versions. To address the problem for lower Julia versions, PkgUtility provides function predownload_artifact to download and unpack the artifact manually for Windows:

predownload_artifact(art_name, artifact_toml)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/artifact/download.jl:46.

PkgUtility.predownload_artifact — Method

What predownload_artifact does are

determine if the artifact exists already

if true, skip the artifact installation

if false

install the artifact using the download_artifact from Pkg.Artifacts

if the installation fails (e.g., for Windows)

download the .tar.gz manually

unpack the .tar.gz manually

Method for this function is

predownload_artifact(art_name::String, artifact_toml::String)

Download the artifact from given server if it does not exist, given

art_name Artifact name to install

artifact_toml Artifacts.toml file location

Examples

# download the artifact manually if default methods fail
predownload_artifact("test_artifact", "Artifacts.toml");

Date

PkgUtility.month_days — Function

month_days(year::Int, month::Int)

How many days in a month, given

year Year

month Month

Examples

# inquire how many days in 2020 October
month_days(2020, 10);

PkgUtility.month_ind — Function

month_ind(year::Int, doy::Int)

Which month does the day of year belongs to, given

year Year

doy Day of year (typically 1-365, 1-366 for leap years)

Examples

# inquire which month it is for 100th day in 2020
month_ind(2020, 100);

PkgUtility.parse_timestamp — Function

Researchers may often run into the scenarios when they have to convert dates to strings or numbers, vice versa. Thought the conversions are easy to make, it is not convenient at all if ones need to recreate the wheel everytime. Thus, we provide functions to make the conversion. One of the common conversions is to convert time stamp to a number:

parse_timestamp(time_stamp; in_format, out_format)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/date/doy.jl:89.

parse_timestamp(year, doy)
parse_timestamp(year, doy, sep)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/date/doy.jl:157.

PkgUtility.parse_timestamp — Method

The time stamp can be either an integer or a string, and the integer will be converted to a string within the function. The supported time stamp formats are YYYYMMDD, YYYYMMDDhh, YYYYMMDDhhmm, and YYYYMMDDhhmmss, where the labels are

YYYY Year number

MM Month number

DD Day number

hh Hour number

mm Minute number

ss second number

The supported outputs are

DATE A Dates.Date type variable

DATETIME A Dates.DateTime type variable

DOY A day of year integer

FDOY A day of year float

parsetimestamp( timestamp::Union{Int,String}; informat::String = "YYYYMMDD", outformat::String = "DOY")

Convert time stamp, given

time_stamp Time stamp

in_format Format of time_stamp, default is YYYYMMDD

out_format Output format, default is DOY

Examples

# if input format is YYYYMMDD
_inf = "YYYYMMDD";
for _ts in [20200130, "20200130"]
    for _outf in ["DATE", "DATETIME", "DOY", "FDOY"]
        parse_timestamp(_ts; in_format=_inf, out_format=_outf);
    end
end

# if input format is YYYYMMDDhh
_inf = "YYYYMMDDhh";
for _ts in [2020013002, "2020013002"]
    for _outf in ["DATE", "DATETIME", "DOY", "FDOY"]
        parse_timestamp(_ts; in_format=_inf, out_format=_outf);
    end
end

# if input format is YYYYMMDDhhmm
_inf = "YYYYMMDDhhmm";
for _ts in [202001300215, "202001300215"]
    for _outf in ["DATE", "DATETIME", "DOY", "FDOY"]
        parse_timestamp(_ts; in_format=_inf, out_format=_outf);
    end
end

# if input format is YYYYMMDDhhss
_inf = "YYYYMMDDhhmmss";
for _ts in [20200130021515, "20200130021515"]
    for _outf in ["DATE", "DATETIME", "DOY", "FDOY"]
        parse_timestamp(_ts; in_format=_inf, out_format=_outf);
    end
end

PkgUtility.parse_timestamp — Function

Also, one might want to convert date to time stamp. Supported conversions are

parse_timestamp(year::Int, doy::Int, sep::String="")

Convert date to YYYY(sep)MM(sep)DD format, given

year Year

doy Day of year (typically 1-365, 1-366 for leap years)

sep String that separates year, month, and day. Default is ""

Examples

# convert the time to YYYYMMDD format
parse_timestamp(2020, 100);
parse_timestamp(2020, 100, "");

# convert the time to YYYY.MM.DD format
parse_timestamp(2020, 100, ".");

Display

PkgUtility.pretty_display — Function

Display the parameters in a colored way.

pretty_display(dict, max_len)
pretty_display(dict, max_len, spaces)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/display/recursive.jl:38.

pretty_display(dicts)
pretty_display(dicts, spaces)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/display/recursive.jl:100.

PkgUtility.pretty_display — Function

When a dictionary (a pair of key and item) is passed to the pretty_display function, the dictionary will be displayed as (LEADING_SPACE)key ⇨ item,. However, if the dictionary item is an array of dictionaries, then the array will be displayed recursively.

pretty_display(dict::Pair, max_len::Int, spaces = "    ")

Display the key and item in a dictionary (Pair), given

dict Dictionary to display

max_len Maximum length of dictionary key (in an array)

spaces Leading spaces before displaying the dictionary key

Examples

# display the pair in a pretty way (automatic alignments)
pretty_display("a" => "b", 1);
pretty_display("a" => "b", 2, "  ");

PkgUtility.pretty_display — Function

When an array of dictionaries is given, the pretty_display function computes the maximum length of the dictionary keys, and then display the dictionaries in a colored and pretty manner

pretty_display(
            dicts::Union{Vector{Pair{String,String}},
                         Vector{Pair{String,Any}},
                         Vector{Pair{Any,String}},
                         Vector{Pair{Any,Any}}},
            spaces::String = "    ")

Display array of pairs (dictionary) in recursive manner, given

dicts Parameters to display

spaces Leading spaces

Examples

# display the vector of pairs recursively (automatic alignments)
_dicts = ["A" => "b", "d" => "A", "rr" => ["ra" => "rB", "rD" => "ra"]];
pretty_display(_dicts);
pretty_display(_dicts, "  ");

Math

Integral function

PkgUtility.numerical∫ — Function

Function numerical∫ calculates the integral numerically. The supported methods are

numerical∫(f, Δx)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/math/integral.jl:39.

numerical∫(f, Δx)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/math/integral.jl:71.

numerical∫(f, x_min, x_max, n)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/math/integral.jl:107.

numerical∫(f, x_min, x_max)
numerical∫(f, x_min, x_max, x_tol)
numerical∫(f, x_min, x_max, x_tol, y_tol)

defined at /home/runner/work/PkgUtility.jl/PkgUtility.jl/src/math/integral.jl:127.

Quadratic solver

PkgUtility.lower_quadratic — Function

lower_quadratic(a::FT, b::FT, c::FT) where {FT<:AbstractFloat}

Return the lower quadratic solution or NaN, given

a Parameter in a*x^2 + b*x + c = 0

b Parameter in a*x^2 + b*x + c = 0

c Parameter in a*x^2 + b*x + c = 0

PkgUtility.upper_quadratic — Function

upper_quadratic(a::FT, b::FT, c::FT) where {FT<:AbstractFloat}

Return the upper quadratic solution or NaN, given

a Parameter in a*x^2 + b*x + c = 0

b Parameter in a*x^2 + b*x + c = 0

c Parameter in a*x^2 + b*x + c = 0

Statistics extensions

PkgUtility.mae — Function

mae(y::Array, pred::Array)

Return the mean absolute error by ommiting the NaN, given

y Array of numbers, can be NaN

pred Array of predictions, can be NaN

PkgUtility.mape — Function

mape(y::Array, pred::Array)

Return the mean absolute percentage error by ommiting the NaN, given

y Array of numbers, can be NaN

pred Array of predictions, can be NaN

PkgUtility.mase — Function

mase(y::Array, pred::Array)

Return the mean absolute standardized error by ommiting the NaN, given

y Array of numbers, can be NaN

pred Array of predictions, can be NaN

PkgUtility.nanmax — Function

nanmax(x::Array)

Return the maximum of array ommiting the NaN, given

x Array of numbers, can be NaN

PkgUtility.nanmean — Function

nanmean(x::Array)

Return the mean of array by ommiting the NaN, given

x Array of numbers, can be NaN

PkgUtility.nanmedian — Function

nanmedian(x::Array)

Return the median of array by ommiting the NaN, given

x Array of numbers, can be NaN

PkgUtility.nanmin — Function

nanmin(x::Array)

Return the maximum of array ommiting the NaN, given

x Array of numbers, can be NaN

PkgUtility.nanstd — Function

nanstd(x::Array)

Return the std of array by ommiting the NaN, given

x Array of numbers, can be NaN

PkgUtility.rmse — Function

rmse(y::Array, pred::Array)

Return the root mean square error by ommiting the NaN, given

y Array of numbers, can be NaN

pred Array of predictions, can be NaN

NetCDF extensions

Missing docstring.

Missing docstring for ncread. Check Documenter's build log for details.

Recursive test

PkgUtility.FT_test — Function

FT_test(para::Any, FT)

Test the the floating point type of para is FT, given

para Any type of parameters

FT given FT

If data type is not supported, use pass!

PkgUtility.NaN_test — Function

NaN_test(para::Any)

Test the the floating point type of para is not NaN, given

para Any type of parameters

If data type is not supported, use pass!