API

Artifacts

PkgUtility.predownload_artifact — Function

predownload_artifact(name::String, artifact_toml::String)

Download the artifact from given server if it does not exist, given

name Artifact name

artifact_toml Artifacts.toml file location

Display

PkgUtility.pretty_display — Function

pretty_display(
            para::Array{Pair{Union{Any},Union{Any}},1},
            spaces::String = " ")

Display array of pairs (dictionary) in recursive manner, given

para Parameters to display

spaces Leading spaces

Recursive test

PkgUtility.FT_test — Function

FT_test(para::Any, FT)

Test the the floating point type of para is FT, given

para Any type of parameters

FT given FT

If data type is not supported, use pass!

PkgUtility.NaN_test — Function

NaN_test(para::Any)

Test the the floating point type of para is not NaN, given

para Any type of parameters

If data type is not supported, use pass!